# apps/kernel/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI-BOS Kernel API
  version: 1.0.0
  description: |
    Control Plane API for AI-BOS Finance platform.
    
    ## Architecture
    The Kernel is the Control Plane that orchestrates:
    - **Authentication:** JWT-based identity verification
    - **Authorization:** RBAC permission enforcement
    - **Gateway:** Reverse proxy routing to Cells
    - **Audit:** Immutable event logging with correlation IDs
    
    ## Authentication
    Most endpoints require JWT Bearer token (obtained via `/api/kernel/iam/login`).
    
    ## Bootstrap
    First user creation requires `x-kernel-bootstrap-key` header (see Cell Integration Guide).
    
    ## Tenant Authority
    - **Before login:** `x-tenant-id` header selects tenant context
    - **After login:** JWT `tid` claim is authoritative; header is ignored
    
    ## Data Persistence
    All state is persisted to PostgreSQL (survives restarts).
    
    ## Reference Cell
    The `cell-payment-hub` demonstrates cell-based resilience:
    - Health endpoints: `/ping`, `/health`
    - Finance endpoints: `/payments/process`, `/payments/status/:id`
    - Chaos endpoints: `/chaos/fail/:cell`, `/chaos/recover/:cell`
servers:
  - url: http://localhost:3001
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
...
  headers:
    XTenantId:
      schema:
        type: string
      description: Tenant ID (required for control-plane calls and public gateway calls)
      required: true
...
    XKernelBootstrapKey:
      schema:
        type: string
      description: Bootstrap key for initial setup (required for bootstrap endpoints only)
      required: false
  schemas:
    Error:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [UNAUTHORIZED, FORBIDDEN, VALIDATION_ERROR, NOT_FOUND, INTERNAL_ERROR, BOOTSTRAP_DENIED]
            message:
              type: string
            details:
              type: array
              items:
                type: object
        correlation_id:
          type: string
    Success:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          example: true
        data:
          type: object
        correlation_id:
          type: string
paths:
  /api/kernel/iam/users:
    post:
      summary: Create user (bootstrap)
      description: |
        Creates a new user. Requires bootstrap key for empty tenants.
        After tenant has users, this endpoint requires RBAC permission `kernel.iam.user.create`.
      operationId: createUser
      parameters:
        - $ref: '#/components/headers/XTenantId'
        - $ref: '#/components/headers/XKernelBootstrapKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name]
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  tenant_id:
                    type: string
                  email:
                    type: string
                  name:
                    type: string
                  status:
                    type: string
                    enum: [ACTIVE, SUSPENDED]
                  created_at:
                    type: string
                    format: date-time
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (bootstrap denied or insufficient permissions)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    get:
      summary: List users
      description: Lists all users for the tenant (requires JWT)
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/headers/XTenantId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        email:
                          type: string
                        name:
                          type: string
                        status:
                          type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/kernel/iam/users/{userId}/set-password:
    post:
      summary: Set user password
      description: |
        Sets password for a user. Requires bootstrap key if user has no credentials.
        Otherwise requires RBAC permission `kernel.iam.credential.set_password`.
      operationId: setPassword
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/headers/XTenantId'
        - $ref: '#/components/headers/XKernelBootstrapKey'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (bootstrap denied or insufficient permissions)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...